fix: add check on install

+++ install.sh
@@ -4,13 +4,16 @@ echo "Install Pythons Deps"
 python3 -m pip install -r requirements.txt
 
 echo "Setting up env"
-cp .env.example .env
+cp -n .env.example .env
 nano ./.env 
 set -e
 
 echo "Setting up Tide Service"
 service_location="/lib/systemd/system/tide.service"
-cat <<EOF > $service_location
+if [ -f "$service_location" ]; then
+    echo "$service_location exists. Skipping Creation"
+else 
+    cat <<EOF > $service_location
 Description=Tidetracker eInk
 After=multi-user.target
 [Service]
@@ -23,6 +26,8 @@ ExecStart=/usr/bin/python3 -m tidetracker.run
 [Install]
 WantedBy=multi-user.target
 EOF
+fi
+
 
 sudo chmod 644 $service_location
 
@@ -34,7 +39,6 @@ sudo systemctl status tide.service
 echo ""
 echo "Setup Complete: To start/stop/etc.. "
 echo "> sudo systemctl start|stop|restart|status tide"
-
 echo ""
 echo "The config for the service lives here $service_location"
 echo ""
